+++ src/main/java/com/spring/test/SpringSecurity_JWT_test/service/WithdrawalService.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Component;
 @Component
 public interface WithdrawalService {
 
-    public Withdrawal saveWithdrawal(Withdrawal withdrawal, Long id);
+    public Withdrawal saveWithdrawal(Withdrawal withdrawal, Integer id);
 }
